requirements.txt
aiogram==2.19
Pillow==9.0.1
torch==1.10.2
torchvision==0.11.3
python-dotenv==0.19.2